Rename extension-sdk to extensions-sdk
While it is never nice to rename an already published package,
this allows us to reserve the "@directus/extension-" prefix for official extensions.
