+++ app/package.json
@@ -28,7 +28,7 @@
 	"gitHead": "24621f3934dc77eb23441331040ed13c676ceffd",
 	"devDependencies": {
 		"@directus/docs": "9.0.0-rc.91",
-		"@directus/extension-sdk": "9.0.0-rc.91",
+		"@directus/extensions-sdk": "9.0.0-rc.91",
 		"@directus/format-title": "9.0.0-rc.91",
 		"@directus/shared": "9.0.0-rc.91",
 		"@fullcalendar/core": "5.9.0",
