+++ docs/guides/modules.md
@@ -128,12 +128,12 @@ npm init -y
 ```
 
 To be read by the Admin App, your custom module's Vue component must first be bundled into a single `index.js` file. We
-recommend bundling your code using the directus-extension CLI from our `@directus/extension-sdk` package. The CLI
+recommend bundling your code using the directus-extension CLI from our `@directus/extensions-sdk` package. The CLI
 internally uses a Rollup configuration tailored specifically to bundling Directus extensions. To install the Extension
 SDK, run this command:
 
 ```bash
-npm i -D @directus/extension-sdk
+npm i -D @directus/extensions-sdk
 ```
 
 For the directus-extension CLI to recognize the extension type, the input path and the output path, add this field to
