+++ packages/create-directus-extension/package.json
@@ -15,7 +15,7 @@
 	"license": "GPL-3.0-only",
 	"gitHead": "24621f3934dc77eb23441331040ed13c676ceffd",
 	"dependencies": {
-		"@directus/extension-sdk": "9.0.0-rc.91",
+		"@directus/extensions-sdk": "9.0.0-rc.91",
 		"@directus/shared": "9.0.0-rc.91",
 		"inquirer": "^8.1.2"
 	}
