+++ packages/extensions-sdk/package.json
@@ -1,5 +1,5 @@
 {
-	"name": "@directus/extension-sdk",
+	"name": "@directus/extensions-sdk",
 	"version": "9.0.0-rc.91",
 	"description": "A toolkit to develop extensions to extend Directus.",
 	"main": "dist/cjs/index.js",
