+++ packages/extensions-sdk/readme.md
@@ -1,3 +1,3 @@
-# @directus/extension-sdk
+# @directus/extensions-sdk
 
 A toolkit to develop extensions to extend Directus.
