+++ packages/extensions-sdk/src/cli/commands/build.ts
