+++ packages/extensions-sdk/src/cli/utils/rename-map.ts
