+++ packages/extensions-sdk/templates/common/_gitignore
