+++ packages/extensions-sdk/templates/display/src/display.vue
