+++ packages/extensions-sdk/templates/display/src/index.js
