+++ packages/extensions-sdk/templates/endpoint/src/index.js
