+++ packages/extensions-sdk/templates/hook/src/index.js
