+++ packages/extensions-sdk/templates/interface/src/index.js
