+++ packages/extensions-sdk/templates/interface/src/interface.vue
