+++ packages/extensions-sdk/templates/layout/src/index.js
