+++ packages/extensions-sdk/templates/module/src/index.js
