+++ packages/extensions-sdk/templates/module/src/module.vue
