+++ packages/shared/src/constants/extensions.ts
@@ -1,4 +1,4 @@
-export const APP_SHARED_DEPS = ['@directus/extension-sdk', 'vue', 'vue-router'];
+export const APP_SHARED_DEPS = ['@directus/extensions-sdk', 'vue', 'vue-router'];
 export const API_SHARED_DEPS = ['axios'];
 
 export const APP_EXTENSION_TYPES = ['interface', 'display', 'layout', 'module'] as const;
